+++ libraries/BURT_arm_motor/BURT_arm_motor.h
@@ -1,7 +1,8 @@
 #ifndef burt_arm_motor_h
 #define burt_arm_motor_h
 
-// #include 
+#include <BURT_arm_constants.h>
+#include <TMC_utils.h>
 
 class Motor { 
 	public: 
@@ -10,7 +11,7 @@ class Motor {
 		 Pass in the motor's limits (in radians) here. The [updateAngle] method
 		 will be sure to never pass these angles.
 	 */
-		Motor(double limits[2]);
+		Motor(const double _limits[2]);
 
 		/* Returns true if the motor stalled recently. */
 		bool didStall();
@@ -22,9 +23,9 @@ class Motor {
 		void safeUpdate(double newAngle);
 
 	private: 
-		TMC5160Stepper driver;
+		// TMC5160Stepper driver = NULL;
 		double angle = 0;
-		double limits[2];
+		double limits[2] = {-1, -1};
 };
 
 #endif
